Breaking: replace `bump` with `cc add` command
Existing behavior remains the same, so `hallmark cc add <target>`
does the same as `hallmark bump <target>` did.

The new command is more generic however: it also supports inserting
a release into the changelog for an existing version (regardless of
whether that version is first, last or somewhere in the middle).

In addition, there's a new option `--no-commits` to skip populating
the release with commits. Effectively only creating a markdown
heading (with a version number, link and date).

Closes https://github.com/vweevers/hallmark/issues/84